Changed DomainObject.java and TrainRide.java. Added date to train ride.

+++ src/main/java/com/tsystems/jschool/railage/domain/DomainObject.java
@@ -9,13 +9,13 @@ public abstract class DomainObject {
 
 
     /** Unique id of DomainObject instance */
-    protected Long id;
+    protected Integer id;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 }
